Consolidated and improved integration tests for register_disable_route.
